[TEST] generator 컨벤션에 맞추어 코드 수정 (#43)

+++ src/main/java/com/debatetimer/service/member/MemberService.java
@@ -20,8 +20,8 @@ public class MemberService {
     private final ParliamentaryTableRepository parliamentaryTableRepository;
 
     @Transactional(readOnly = true)
-    public TableResponses getTables(long memberId) {
-        Member member = memberRepository.getById(memberId);
+    public TableResponses getTables(Long memberId) {
+        Member member = memberRepository.getById(memberId.longValue());
         List<ParliamentaryTable> parliamentaryTable = parliamentaryTableRepository.findAllByMember(member);
         return TableResponses.from(parliamentaryTable);
     }
